+++ capstone_flutter/lib/view/screen/pulsa&paket_data_screen/metode_pembayaran_pulsa_screen.dart
@@ -328,8 +328,8 @@ class _MetodePembayaranPulsaScreenState
                       provider: widget.provider,
                       description: widget.description,
                       createdAt: widget.createdAt,
-                      balanceNow: int.parse(myBalance.toString()) -
-                          int.parse(widget.price),
+                      balanceNow: double.parse(myBalance.toString()) -
+                          double.parse(widget.price),
                     ),
                   ),
                 );
