+++ capstone_flutter/lib/view/screen/pulsa&paket_data_screen/paket_data/metode_pembayaran_paket_data_screen.dart
@@ -334,8 +334,8 @@ class _MetodePembayaranPaketDataScreenState
                       provider: widget.provider,
                       description: widget.description,
                       createdAt: widget.createdAt,
-                      balanceNow: int.parse(myBalance.toString()) -
-                          int.parse(widget.price),
+                      balanceNow: double.parse(myBalance.toString()) -
+                          double.parse(widget.price),
                     ),
                   ),
                 );
