+++ capstone_flutter/lib/view/screen/pulsa&paket_data_screen/paket_data/pin_paket_data_screen.dart
@@ -17,7 +17,7 @@ class PinPaketDataScreen extends StatefulWidget {
   final String price;
   final String adminFee;
   final String description;
-  final int balanceNow;
+  final double balanceNow;
   final DateTime createdAt;
   final String token;
