+++ capstone_flutter/lib/view/screen/pulsa&paket_data_screen/pin_pulsa_screen.dart
@@ -18,7 +18,7 @@ class PinPulsaScreen extends StatefulWidget {
   final String price;
   final String adminFee;
   final String description;
-  final int balanceNow;
+  final double balanceNow;
   final DateTime createdAt;
   final String token;
